Update SyDNEy.py
Added #mix coordinator for applying different instructions, models, etc. to a latent, step by step, and mixing the results at each step.  The default is to mix the results evenly, but an $at setting can be used to specify a fraction for a given #mix (for example, #mix ... $at 0.7).  A #mix coordinator must come after at least one #step coordinator specifying something for it to mix WITH.